add simple html dom parser to codeignitor as a application/helpers: how to make it work: http://ellislab.com/forums/viewthread/114059/
Waldmeister is the master. Thanks for your persistence in getting me to try it as a helper, rather than a library, as my lack in understanding the subtle differences between a CI Library and CI Helper seems to have been the main problem for me getting it to work.

I got it to work by doing the following steps:

1) Drop the file simple_html_dom.php into /application/helpers folder
2) Change the name to simple_html_dom_helper.php
3) Capitalize the first letter of the two class names in the file:
around line 83 class Simple_html_dom_node
around line 481 class Simple_html_dom
4) Then remember to load the new helper in the controller with $this->load->helper(‘simple_html_dom’)